+++ lib/selectors/list.ts
@@ -6,6 +6,7 @@ export type List<T extends object> = {
 	type: "LIST";
 	convert: Array<T>;
 	// ---Additional---
+	predicate?: (el: Cheerio, index: number) => boolean;
 	selector: string;
 	data: Query | Selector;
 };
@@ -21,10 +22,20 @@ export const listResolve = <Q extends object>(
 	const els = $(queryType.selector, context);
 	for (let i = 0; i < els.length; i++) {
 		const el = els.eq(i);
-		const scrapedData = isSelector(queryType.data)
-			? scrapSelector($, el, queryType.data)
-			: scrapQuery($, el, queryType.data, {});
-		result.push(scrapedData);
+		if (queryType.predicate) {
+			if (queryType.predicate(el, i)) {
+				const scrapedData = isSelector(queryType.data)
+					? scrapSelector($, el, queryType.data)
+					: scrapQuery($, el, queryType.data, {});
+				result.push(scrapedData);
+			}
+		} else {
+			const scrapedData = isSelector(queryType.data)
+				? scrapSelector($, el, queryType.data)
+				: scrapQuery($, el, queryType.data, {});
+			result.push(scrapedData);
+		}
+
 	}
 	return result;
 };
@@ -33,13 +44,15 @@ export const listResolve = <Q extends object>(
  * Get list of items
  * @param selector - css selector for list of items
  * @param data - query per item
+ * @param predicate - filter elements
  */
 export const listCreator = <Q extends Query | Selector>(
 	selector: string,
-	data: Q
+	data: Q,
+	predicate?: (el: Cheerio, index: number) => boolean
 ): List<Q extends Query
 	? TypeOfQuery<Q>
 	: Q extends Selector
 		? TypeOfSelector<Q>
 		: never
-> => ({ type: "LIST", convert: [], selector, data });
+> => ({ type: "LIST", convert: [], selector, data, predicate });