+++ lib/selectors/select.ts
@@ -19,8 +19,13 @@ export const selectResolve = <C extends (el: Cheerio) => any>(
 	}
 };
 
-export const selectCreator = <C extends (el: Cheerio) => any>(selector: string, convert: C): Select<C> => ({
+/**
+ * Create custom selector
+ * @param selector - css selector
+ * @param callback - callback with cheerio element
+ */
+export const selectCreator = <C extends (el: Cheerio) => any>(selector: string, callback: C): Select<C> => ({
 	type: "SELECT",
-	convert: convert,
+	convert: callback,
 	selector: selector
 });